+++ crates/blockchain-tree/src/lib.rs
@@ -32,4 +32,6 @@ pub use post_state_data::{PostStateData, PostStateDataRef};
 
 /// Buffer of not executed blocks.
 pub mod block_buffer;
+mod canonical_chain;
+
 pub use block_buffer::BlockBuffer;
